# [v2] performance: Improve ParseSigned whitespace stripping