--- a/signing_test.go
+++ b/signing_test.go
@@ -570,3 +570,13 @@ func TestSignerB64(t *testing.T) {
 		t.Errorf("Input/output do not match, got '%s', expected '%s'", output, input)
 	}
 }
+
+func BenchmarkParseSigned(b *testing.B) {
+	msg := `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c`
+	for i := 0; i < b.N; i++ {
+		_, err := ParseSigned(msg)
+		if err != nil {
+			b.Errorf("Error on parse: %s", err)
+		}
+	}
+}