- When profiling some other code, I noticed ParseSigned wasn't as fast as I expected.
- Adds a benchmark for ParseSigned
- Replaces regexp to remove whitespace with a string.Builder approach.

`benchcmp` results for `go test -v -run=NONE -bench BenchmarkParseSigned -benchtime=30s -benchmem`
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkParseSigned-4     8184          5461          -33.27%

benchmark                  old allocs     new allocs     delta
BenchmarkParseSigned-4     50             48             -4.00%

benchmark                  old bytes     new bytes     delta
BenchmarkParseSigned-4     3832          3656          -4.59%
```